add example script for zae-engine
